+++ src/main/kotlin/com/digitalsubstrate/dsm/liveTemplates/DSMTemplateContextType.kt
@@ -5,8 +5,7 @@ import com.digitalsubstrate.dsm.core.DSMLanguage
 import com.intellij.codeInsight.template.TemplateActionContext
 import com.intellij.codeInsight.template.TemplateContextType
 
-@Suppress("DEPRECATION")
-class DSMTemplateContextType : TemplateContextType("DSM", "DSM") {
+class DSMTemplateContextType : TemplateContextType("DSM") {
     override fun isInContext(templateActionContext: TemplateActionContext): Boolean {
         return templateActionContext.file.language == DSMLanguage.INSTANCE
     }